Add Basic RBAC with CanCanCan
This commit adds basic Role Based Access Control with CanCanCan. It adds
a role class which contains the CanCanCan ability definitions as well as
a name so that each role can be listed and assigned to users more
easily. A User#generate_abilities method has been added to iterate over
all of a User's roles and call Role#apply_abilities on each to create
the users access profile.

This commit also adds appropriate factories for roles and a couple of
different types of users that auto populate roles. Most notable being an
admin_user that is best used for testing user functions that are not
concerned with access control. This user has the :manage, :all arguments
to the Ability#can method.