Add CanCanCan support and Utilize Session model
This commit incorporates CanCanCan methods into controllers and starts
using the Session model logic instead of just a session controller. This
allows use to login users without having to hit the session controller.
This is very useful if we want to log a user in after a certain activity
(password reset) for instance.

This commit also updates some of the controller specs to use an admin
user whom has been given admin rights through the CanCanCan gem. This
allows us to not worry about permissions on tests that are not concerned
with permissions.