Remove Survey/results resource
It was initially thought that all surveys would be dynamically generated
and a results resource would be possible for each survey. I'm not so
sure about that now. I think the best thing to do is have meta data
about each survey created as active record resources that point to new
surveys that are not generated dynamically. These surveys will then have
their results under the show/index actions of said new survey. This way
the programmer has the freedom to create whatever survey is needed. This
will cause the need to redeploy for each new survey, but I think that is
acceptable. For this reason, I'm removing the nested route
surveys/results.