Add User to Contact Controller spec
The first two tests from the contacts_controller_spec check for proper
rendering of the contacts index. We recently added an authorization call
on the Contacts controller. This was causing failing specs. A user with
a contacts_admin_role was created before each of these tests to fix the
issue.