Add UdlModuleSection Position functionality
The previous way the UdlModuleSection.position attribute was implemented was not
going to work for us. This is because we were storing the position on
the UdlModuleSection Object. Turns out, we need a many-to-many
relationship with modules and sections. Some boiler plate sections (i.e.
Introduction sections) will belong to multiple modules and those modules
may need the option of re-ordering sections to accommodate the author's needs.
Therefore, the position attribute has been moved to the association
object between the UdlModule and it's sections. Because a section will
not belong to a module more than once, this is okay.