Remove UdlModuleSectionsController#index and assoc. routing
I'm not sure why I thought I needed an index of sections before, but
there is no need for them with our current layout. The sections all
display an edit button inside the module and we will display links to
the shared sections and template sections on the module index page.
Therefore, this commit removes the associated files with the
UdlModuleSectionsController#index action.