Add shared and default_shared_position to udl_module_sections
The udl module sections will have shared sections, such as the common
introduction sections, the udl principles sections, the literature base
section, and the references and resources sections. These will be normal
sections with the shared boolean set to true. Also, the
default_shared_position will be used to set the initial position on the
module/section association which can then be altered at a later date.