Fix UdlModuleSectionsController#update bug
The UdlModuleSectionsController#update method was not saving because
there was a mismatch in names for the params hash. In a previous commit
we changed the method name for the strong params from section_params to
udl_module_section_params because this is the default that cancan looks
for to authorize a resource. In doing so, I forgot to update the
UdlModuleSection#update call with the new method name. I also added a
test to catch any regressions in UdlModuleSectionsController#update in
the future.