Scope UdlModuleSections in UdlModuleSectionsController
We were only searching for section slugs when setting the
udl_module_section in the edit, update, and delete actions of the
UdlMoudleSectionsController. This introduced a bug when two sections
from different modules had the same name. So we've created a finder
method on UdlModuleSections that takes a module_slug param and a
section_slug param to find the correct section for a given module.