Use sass import in application.css.scss
I've just discovered that I can use @import instead of sprockets
//= require to include files into the asset pipeline in
application.css.scss. This follows the convention I use elsewhere in the
codebase, so I'm using it here. This also negates the need for
app/assests/stylesheets/main.css.scss, so I've move it's contents to
application.css.scss and removed it.