Add initial Redesign Summary Resource
This initial version of the redesign summary resource is functional, but
it still needs some work. Most notably, the file upload is a little
weak. We need to do a better job of informing the user on the status of
attachment uploads. We also have no authorization logic on the views.