Update PasswordResetMailer to use sendgrid
PasswordResetMailer was using mandrill which now requires a paid
mailchimp account. So we are migrating to Sendgrid. We were also using
mandrill templates so the switch to sendgrid will also include using
rails mailer views. If in the future we need to delegate email views to
sendgrid templates that is still a possibility.