Remove class from reference p in udl_module sass
The reference div defines the references section of modules. We had a
class on each paragraph in this section denoting it as a reference.
We'll only ever have references in these sections so it doesn't make
sense to have to put a class on each reference paragraph. Therefore, we
defined a paragraph class for references sections that doesn't have to
define the class.