Add column height fix for sidebar (#6)
* Add Helper spec for UdlModuleHelper

Adds a helper spec for the UdlModuleHelper with tests around the
module_admin? method. This method takes a user as an argument and
determines whether or not a user has module_admin privileges.

* Add column height fix for sidebar

The side bar on the main marketing pages for collegestar.org needs to
get its height information from the height of the main content section.
app/assets/javascripts/column_fix.js does this.