Add redirect routes for legacy.collegestar.org
The naming scheme for the udl modules changed from a code representing
the title of the modules to a slugged title. We still need to route
those old urls to the correct place. Some of the new modules are not
ready yet, so we will 301 redirect them to the legacy.collegestar.org
site until they are live on collegestar.org.