Update controller specs for rails 5
Rails 5 controller specs have under gone a lot of changes. We needed to
add a params: {} around the parameters of each request call. Also, we've
removed the application_controller_spec because it was causing failing
tests and after reviewing what was being tested, those features should
be tested with feature specs, not controller specs.