Update udl_modules_controller_spec
current_user call was failing in a spec because we were not deleting the
auth_token cookie, but instead were just setting it to nil. In rails 5
this was causing an issue, so we've switched to a correct way of
handling the removal of a hash key/value: hash.delete :key.