Remove foundation-rails gem
We moved to using bourbon instead of foundation. The reasoning is we
were using foundation not as a framework, but as a library. This worked
great until we tried upgrading foundation and realized that we had
basically been using a private api and it all blew up on upgrade.
Bourbon seems like a better fit for how we'll be using sass frameworks
in the future.