Add view_admin? method to ApplicationPolicy
We have numerous places in our views where we need to list admin links
to different resources. Sometimes the logic that decides if a user can
see these links does not align with any of the crud verb methods that
are used by default in pundit. In those cases, we have add the
view_admin? method to control the display of these links.