Fix logic error in db/seeds
db/seeds.rb erroneously checked for the current column data being equal
to the seed data before it tried to update the current column. This is
the opposite of what is needed. It now checks for differences in the
seed data and the current column data and updates the current info to
match the seed.