Add role assignment to user profile editor
The user's profile editor now allows those with the granter role to add
certain roles to users. These roles are denoted as "grantable". Roles
denoted "non_grantable" will not be loaded into the user's profile view
and must be granted from the command line. At this point, only three of
the roles are non_grantable:

  - Site Administrator
  - Granter
  - User Administrator