Add start date to Module Proposals
Adds a date picker field for a start date to the module proposals
resource. This commit does not set a null: false constraint on this
field because we currently have existing module proposals. After those
proposals have had a default value of '1900-01-01' to their start dates,
we'll push another commit with a null: false constraint migration for
the start_date field.

Also adds a change to the application mailer which allows us to set a
'pretty' from variable in our emails.