Add STAR Learning Community Registration
The STAR Learning Community Registration and SLC landing pages have been
added as well as basic specs for the registration CRUD. Feature specs
are still needed as well as adding a way for SLC community admins to
receive notification when a registration has been submitted. At this
point, Ruben's email address has been hard coded into the response
email. We need to add a way for the admins to subscribe to this event
instead.