Update sessions controller to redirect on login
If a user is sent to the login view because they failed an authorize
call, we want to remember where they were trying to go so we can try to
send them there again after login. If they fail the authorization after
being logged in, we handle that in the resource controller and redirect
from there.