Update UDL Module Index
The udl module index up to this point had been a basic admin index of
all available modules. Moving forward this will be the public index of
the modules. Therefore, we've put the unreleased modules behind an admin
check and also created a modules page content that will allow dynamic
editing of the copy on this page.