Update User email
A bug was discovered in user/session creation. The email address of a
user was being treated case sensitive. This would lead to users not
being able to login if they accidentally created an account with a
capitalized email address. (The person that discovered this used their
phone to create their account and it auto capitalized the first letter
of her email address). Therefore, we're now downcasing user's emails on
create and on session creation so that the email addresses are now case
insensitive.