Update FacultyNominationSurvey faculty reference
A future iteration of this survey will pull in all the currently known
faculty for an institution and provide a list for the user to select the
faculty member from. This has not been released yet, but the database
already has a foreign key/non null constraint on InstitutionalFaculty.
Therefore, we need to assign a value to that reference. We're just
pulling out the first faculty member and assigning them.