Fix bug in udl module section display
After adding a udl alignment section, it was not being displayed in the
UDLModule#show view. This was because we had a mismatch between the
section look up and what we were naming the sections' parent causing it
to fail to load.

To fix this problem from happening again we've added a validation that
checks that a section cannot have a parent that is not present in the
UdlModule::PAGES array. We also added capybara specs testing the adding
of each section type in a module.