Add UserModuleAssessment and UserAssessmentSelectedAnswerChoice
When a user submits a module assessment it will create a new
UserModuleAssessment. We'll use UserAssessmentSelectedAnswerChoice to
track which answer choice they selected. This way we are not tying
ourselves to a fixed number of questions per assessment. This will also
give us the freedom to have multiple answer questions in the future.