Add a SendgridEventPost object
SendGrid send its events in batches therefore, we can't handle each post
as a resource and instead need to store the batch and fire off a delayed
job to create the events. This iteration only saves the post to a
SendgridEventPost object. We'll create the delayed job to create the
events in a later commit.