Add working SMTPAPI headers
We send this email through sidekiq and in the process of trouble
shooting this email, I discovered that we need to be rebooting sidekiq
on deploys and we haven't been. That is why my changes were not working,
sidekiq still had an old version of the app running in memory.