Remove direct upload from 321 video edit
The direct upload feature of active storage is not working for the 321
videos. The first problem is that the upload link is not using https and
causing a mixed media failure. I could find no way to remedy this so
we'll remove it and see if we can get a working solution.