Add media queries for iframe sizes
This was done for youtube video embeds. We raised the default iframe
size of the youtube videos from 560x315 to 800x450 and added media
queries for smaller sizes. We also updated the media query breakpoint
for medium from 500px to 600px. The 500px threshold was never hit
because the container minimums would kick in first.